CC/requirements.txt
Flask
gunicorn
mysql.connector
pandas